Rename pick_one_from_weighted_list to select_from_weighted
