+++ src/lib.rs
@@ -14,6 +14,7 @@ mod js;
 mod pick;
 mod prng;
 mod proxy;
+mod select_from_weighted;
 mod shuffle;
 mod sub_randomness;
 
@@ -22,7 +23,8 @@ pub use decimal::random_decimal;
 pub use dice::roll_dice;
 pub use encoding::{randomness_from_str, RandomnessFromStrErr};
 pub use integers::{int_in_range, ints_in_range, Int};
-pub use pick::{pick, pick_one_from_weighted_list};
+pub use pick::pick;
 pub use proxy::{NoisCallback, ProxyExecuteMsg, ReceiverExecuteMsg, MAX_JOB_ID_LEN};
+pub use select_from_weighted::select_from_weighted;
 pub use shuffle::shuffle;
 pub use sub_randomness::{sub_randomness, sub_randomness_with_key, SubRandomnessProvider};
